Rename endpoints for clarity, add delete index & add/delete document endpoints
